requirements.txt
certifi==2017.1.23
flask==0.12
flask-babel==0.11.1
lxml==3.7.3
idna==2.5
pygments==2.1.3
pyopenssl==16.2.0
python-dateutil==2.6.0
pyyaml==3.12
deepmerge==0.0.2
requests[socks]==2.13.0
